// src/models/INhanKhau.ts
export interface INhanKhau {
    id: string;
    maNhanKhau: string;
    hoTen: string;
    bietDanh: string;
    namSinh: Date;
    gioiTinh: string;
    noiSinh: string;
    nguyenQuan: string;
    danToc: string;
    tonGiao: string;
    quocTich: string;
    soHoChieu: string;
    noiThuongTru: string;
    diaChiHienNay: string;
    trinhDoHocVan: string;
    trinhDoChuyenMon: string;
    bietTiengDanToc: string;
    trinhDoNgoaiNgu: string;
    ngheNghiep: string;
    noiLamViec: string;
    tienAn: string;
    ngayChuyenDen: Date;
    lyDoChuyenDen: string;
    ngayChuyenDi: Date;
    lyDoChuyenDi: string;
    diaChiMoi: string;
    ngayTao: Date;
    idNguoiTao: string;
    ngayXoa: Date;
    idNguoiXoa: string;
    lyDoXoa: string;
    ghiChu: string;
}

export class NhanKhauModel implements INhanKhau {
    id = "";
    maNhanKhau = "";
    hoTen = "";
    bietDanh = "";
    namSinh = new Date(0);
    gioiTinh = "";
    noiSinh = "";
    nguyenQuan = "";
    danToc = "";
    tonGiao = "";
    quocTich = "";
    soHoChieu = "";
    noiThuongTru = "";
    diaChiHienNay = "";
    trinhDoHocVan = "";
    trinhDoChuyenMon = "";
    bietTiengDanToc = "";
    trinhDoNgoaiNgu = "";
    ngheNghiep = "";
    noiLamViec = "";
    tienAn = "";
    ngayChuyenDen = new Date(0);
    lyDoChuyenDen = "";
    ngayChuyenDi = new Date(0);
    lyDoChuyenDi = "";
    diaChiMoi = "";
    ngayTao = new Date(0);
    idNguoiTao = "";
    ngayXoa = new Date(0);
    idNguoiXoa = "";
    lyDoXoa = "";
    ghiChu = "";
}